+++ hegel-macros/src/struct_gen.rs
@@ -5,7 +5,7 @@ use syn::{DeriveInput, Fields};
 use crate::utils::{cbor_to_iter, default_gen_bounds, tuple_schema};
 
 /// Derive Generator for a struct.
-pub(crate) fn derive_struct_generate(input: &DeriveInput, data: &syn::DataStruct) -> TokenStream {
+pub(crate) fn derive_struct_generator(input: &DeriveInput, data: &syn::DataStruct) -> TokenStream {
     let name = &input.ident;
     let generator_name = format_ident!("{}Generator", name);
 
@@ -30,11 +30,8 @@ pub(crate) fn derive_struct_generate(input: &DeriveInput, data: &syn::DataStruct
 
     let field_types: Vec<_> = fields.iter().map(|f| &f.ty).collect();
 
-    // Generate the with_* method names
-    let with_methods: Vec<_> = field_names
-        .iter()
-        .map(|name| format_ident!("with_{}", name))
-        .collect();
+    // Generate the builder method names (same as field names, no prefix)
+    let builder_methods: Vec<_> = field_names.to_vec();
 
     // Generate field definitions for the generator struct
     let generator_fields = field_names
@@ -60,11 +57,11 @@ pub(crate) fn derive_struct_generate(input: &DeriveInput, data: &syn::DataStruct
     // Generator Default trait bounds for new()
     let default_bounds = default_gen_bounds(&field_types, quote! { 'a });
 
-    // Generate with_* methods
+    // Generate builder methods
     let with_method_impls = field_names
         .iter()
         .zip(field_types.iter())
-        .zip(with_methods.iter())
+        .zip(builder_methods.iter())
         .map(|((field_name, field_type), method_name)| {
             quote! {
                 /// Set a custom generator for this field.
@@ -126,6 +123,8 @@ pub(crate) fn derive_struct_generate(input: &DeriveInput, data: &syn::DataStruct
 
     let expanded = quote! {
         const _: () = {
+            use hegel::generators::Generator as _;
+
             pub struct #generator_name<'a> {
                 #(#generator_fields,)*
             }
@@ -155,7 +154,6 @@ pub(crate) fn derive_struct_generate(input: &DeriveInput, data: &syn::DataStruct
 
             impl<'a> hegel::generators::Generator<#name> for #generator_name<'a> {
                 fn do_draw(&self, __data: &hegel::TestCase) -> #name {
-                    use hegel::generators::Generator;
                     if let Some(basic) = self.as_basic() {
                         basic.parse_raw(hegel::generate_raw(__data, basic.schema()))
                     } else {
@@ -169,8 +167,6 @@ pub(crate) fn derive_struct_generate(input: &DeriveInput, data: &syn::DataStruct
                 }
 
                 fn as_basic(&self) -> Option<hegel::generators::BasicGenerator<'_, #name>> {
-                    use hegel::generators::Generator;
-
                     #(#basic_bindings)*
 
                     let schema = #schema_ts;